Pull additional parsing & validation into args.js
