+++ lib/args.js
@@ -1,4 +1,8 @@
-var args = require('yargs')
+var documentation = require('../'),
+  path = require('path'),
+  loadConfig = require('../lib/load_config.js');
+
+var yargs = require('yargs')
   .usage('Usage: $0 <command> [options]')
 
   .option('f', {
@@ -49,4 +53,74 @@ var args = require('yargs')
 
   .example('$0 foo.js', 'parse documentation in a given file');
 
-module.exports = args;
+/**
+ * Parse and validate command-line options for documentation.
+ * @param {Array} args The array of arguments to parse; e.g. process.argv.slice(2).
+ * @return {object} {inputs, options, formatter, formatterOptions, output}
+ * @private
+ */
+module.exports = function (args) {
+  var argv = yargs.parse(args);
+
+  var inputs,
+    name = argv.name,
+    version = argv.version,
+    transform;
+
+  if (argv._.length > 0) {
+    inputs = argv._;
+  } else {
+    try {
+      var p = require(path.resolve('package.json'));
+      inputs = [p.main];
+      name = name || p.name;
+      version = version || p.version;
+      if (p.browserify && p.browserify.transform) {
+        transform = p.browserify.transform;
+      }
+    } catch (e) {
+      yargs.showHelp();
+      throw new Error('documentation was given no files and was not run in a module directory');
+    }
+  }
+
+  if (!documentation.formats[argv.f]) {
+    yargs.showHelp();
+    throw new Error('Formatter not found');
+  }
+
+  if (argv.f === 'html' && argv.o === 'stdout') {
+    yargs.showHelp();
+    throw new Error('The HTML output mode requires a destination directory set with -o');
+  }
+
+  var formatterOptions = {
+  };
+
+  var config = {};
+
+  if (argv.config) {
+    config = loadConfig(argv.config);
+  }
+
+  return {
+    inputs: inputs,
+    options: {
+      private: argv.private,
+      transform: transform,
+      github: argv.github,
+      polyglot: argv.polyglot,
+      order: config.order || [],
+      shallow: argv.shallow
+    },
+    formatter: argv.f,
+    formatterOptions: {
+      name: name,
+      version: version,
+      theme: argv.theme
+    },
+    output: argv.o
+  }
+}
+
+
